Inject dependencies into TokenVerifier
This commit makes the JwkProvider a constructor parameter so that we can mock it later.

(We're overloading the constructor, though, so you can still call it with the old signature and everything's fine.) 

Also, this commit does a bit of refactoring—for example, it puts the Auth0 tenant URL into a constant, so that we can easily change it later.