Test permanently deleting notes
This method has most of the same problems that deleteNote() had. (For example, it currently expects a text/plain response rather than an application/json response.)

In fact, this method is even worse than deleteNote(), because the corresponding endpoint on the server is all commented out. It's really unclear what the API is supposed to be here—I'm making the assumption that the API is exactly analogous to deleteNote() and that *both* the client and the server are in error.

At the moment, two of these tests are passing; two are failing.