Test permanently deleting notes on the server
This test adds on to the existing test for permanently deleting notes. Now, we also check that we're sending back a JSON with the ID of the deleted note in it. (Or giving a 404 Not Found if the note wasn't in the database.)

Three of these tests are currently failing.