Test that we can verify a token that was just issued this very second
To make sure that there aren't any rounding errors in TokenVerifier (along the lines of "we tried to verify the token too early".)

This test currently passes, probably because of the one-second delay in TokenVerifier. Going forward, we're going to try to remove the one second delay, make this test fail, and then add a leeway to TokenVerifier, making this test pass.