Put the verifyTheTokenIssuedRightThisSecond test in a loop
To try to ameliorate its nondeterministicness.

Some experimentation has shown that GitHub actions might be just barely fast enough to produce the failure case for this test, especially if you run the test over and over again. That being said, I still wouldn't trust it with a ten-foot pole.
foo