Rewrite AddNoteComponent using .pipe() instead of .subscribe()
For fun, mostly, but also because it should be easier to test now. Previously, were launching a lot of `subscribe` callbacks, and the component was in an invalid state while it was waiting for those callbacks to complete.

This isn't really a problem in the browser, because humans are slow and the component finishes initializing before they do anything.

But Karma is fast, and we would have to tell Karma, 'hey, this component isn't going to be ready for a while, wait a bit.'

After this commit, there isn't really a concept of an invalid state; if the component isn't ready yet, it'll just wait.