Test the form validation in AddNoteComponent
These tests all pass already.

Note that we aren't explicitly checking the validation messages because some of them overlap. (In particular the `minlength` rule overlaps with the `required` rule; it's not always predictable of the two will trigger, so you can't assert that one or the other message will necessarily be present.)