Test submitting the form
These tests all pass already.

They're structured a bit strangely because the .submitForm() method subscribes to an observable, and we have to wait for that subscription to complete before we can ask it any questions. To do this, we put the call to .submitForm() method in an async before-each clause; then, all of the subscriptions in .submitForm() will complete before we move on to the actual tests. It's a bit of a hack, but it works.