Refactor the submitForm() tests in add-note.component.spec.ts
Now, they use Angular's fixture.whenStable() to wait for submitForm() to complete.

Previously, we had put submitForm() in its own async beforeEach, which accomplished the same task, but made spying on the SnackBar more awkward.

Now, we can spy on the SnackBar on a per-test basis, rather than having one spy that we keep resetting.