Test submitting the form
These tests are very similar the tests from commit 122d46. These tests are currently failing, probably for async-related reasons (we're calling subscribe in ngOnInit, which takes a while to go through.)

This commit also adds a reset() method to MockNoteService, to make sure that we aren't just reading garbage data from previous tests. (Really, we should be using Jasmine's spies for this; I'll make that change in a little bit.)