Fix sharing observables in OwnerComponent
Fixing problems with 9b21167a5aa20a9230baff1c03035a13836654f6.

Essentially, we were sharing the wrong observables.

Previously, we were sharing the observables that OwnerService makes, but this was useless, because each subscription to this.owner or this.notes makes a new call to OwnerService. It doesn't matter if these observables were shared; each subscriber was looking at a different observable.

The fix is to make this.owner and this.notes shared. Then, all subscribers really are looking at the same observable, and each observable only runs once for all the subscribers.