Update AddNotComponent and MockAuthService to reflect changes to AuthService
Notably, we're using getUser$() rather than userProfile$ now.

Some tests for AddNoteComponent are currently failing—I think the problems are RxJS related, but I don't know specifically what they are at the moment. (When I run the code manually, it seems to work fine, so I think the problems are with the tests themselves.)