Increase Javalin's cache size for requests.
Usually, Javalin doesn't want to store large requests in memory, which is a problem if we want to post a really long note. So, we increase the maximum request size that Javalin will cache.

This pull request makes it so that we take the maximum note body length, multiply it by three for good measure, and compare that to the current cache size. If the cache size is too small, we make it bigger.

We could probably also rewrite the NoteController so that we don't need to cache the request ¯\_(ツ)_/¯ but that sounds like a lot of work for an optimization that probably isn't necessary in the long run.