Fix add-note function not working in staging
It turns out that retrieveOwner() in the add-note-component was returning undefined  for the owner because the subscription for the x500 we were grabbing got completed after the program had already called the getOwnerByx500. I've marked the function as async and added an await to the subscription and that seems to have fixed things. 

Completes issue #32