Fixes the trash can not pulling up the owner notes
This ended up being the same issue as the add note component; the x500 subscription completed after the program had already called the function that wanted to use it.

This also includes the add-note fix simply because it hadn't gotten merged into staging yet and it's hard to test a trash can if you don't have notes to delete.

Closes issue #35