Generate a door sign on the server-side
Rather than generating the PDF with JavaScript, which was flaky and
didn't work on all browsers.

As of this commit, the sign-generation endpoint is all set up and
working, but the client-side doesn't make use of it. Also, there's no
JUnit testing. Those two things are next on the list.

Addresses #13.