Add DeathTimer and update NoteController to work with altered types
Change the expireDate to a Date, to match the client and improve ease of use; change the status to an enumeration, replacing the boolean "posted" to allow for drafts in the future; and improve editing notes in order to allow the expiration date and status to be edited.  For issue #11.