Fix malformed notes in the JUnit tests
Sorry, this was my fault from a long time ago—I forgot to add `posted` fields to these notes. (`status` fields, now.) Because we didn't do anything with these notes, it didn't cause the test to fail until DeathTimer started checking for the status field, which didn't exist.

This causes a handful of unit tests to pass.