Remove extra before server calls for verification
I feel like the before calls under get a single note should cover most of the other methods because they all use the api/notes/:id routing, and while the edit note routing is different, it calls getNoteByID and hits the other routing anyway. I'm committing this as a single file change so it'll be easier to change this if I messed something up.

#19