Make AppComponent get the user JIT instead of in ngOnInit()
If AppComponent only reads the user's fields when necessary, we'll be able to use AppComponent when the user is null (that is, when the user is not logged in).

By contrast, if AppComponent tries to read the user's fields in its ngOnInit method, it will immediately throw an error if the user isn't logged in.

(We could also use the safe navigation operator `?.`, in ngOnInit() to produce roughly the same behavior. ngOnInit() wouldn't throw any null errors, but the rest of the methods still might.

All else being equal, I would prefer to keep the null propagation to a minimum, so I'm avoiding the safe navigation operator in favor of not-reading-those-fields-right-now.)